chore(#373): rename internal directory
Renamed internal directory for dataset-
endpoints, to be more consistent to the rest.
