+++ src/hanami/src/api/http/dataset/delete_dataset_v1_0.h
