+++ src/hanami/src/api/http/dataset/mnist/create_mnist_dataset_v1_0.cpp
