+++ src/hanami/src/api/http/dataset/mnist/finalize_mnist_dataset_v1_0.cpp
